refactor: add computer repository class and model class
